+++ packages/datastore/datastore/src/utils/logger.ts
@@ -1,11 +1,11 @@
 import debug, { Debugger } from "debug";
 
 export function createLogger(name: string): Debugger {
-  return debug(`datasync:${name}`);
+  return debug(`datastore:${name}`);
 }
 
 export function enableLogger() {
   if (window?.localStorage) {
-    window.localStorage.setItem("debug", "datasync:*");
+    window.localStorage.setItem("debug", "datastore:*");
   }
 }
